Fix offscreen rendering logic

+++ cel_app/src/terminal_widget.rs
@@ -241,13 +241,13 @@ impl TerminalWidget {
         default_height: f32,
         bg_color: &[f32; 3]
     ) {
-        let mut line_offset = 0.0;
+        let mut line_offset = 0;
         let mut padding_px = self.padding_px;
 
         // If the alt screen buf is active, we can ignore special rendering styles
         // Also, ensure the most recent screen is visible (home cursor)
         if self.ansi_handler.is_alt_screen_buf_active() {
-            line_offset = self.ansi_handler.get_terminal_state().global_cursor_home[1] as f32;
+            line_offset = self.ansi_handler.get_terminal_state().global_cursor_home[1] as u32;
             padding_px = [0.0, 0.0];
         }
